toolchain: close out correctness-audit P0 items; linker now fails on bad relocations
Real fixes:
- s32-ld: out-of-range/misaligned BRANCH/JAL/CALL and unpaired LO12
  relocation errors now count into reloc_errors and fail the link
  (previously printed to stderr, then exited 0 with a corrupt executable)
- slow32asm: .half/.byte with a symbol operand now error out via
  parse_data_immediate() instead of silently emitting the addend (0)
- slow32dump: bounds checks moved to 64-bit math; str_offset+str_size
  (s32o and s32x) and sec->offset+sec->size could wrap and bypass guards

Docs/tests:
- INSTRUCTION-SET.md: `not` pseudo row now documents the actual
  sub+addi expansion (old row showed the broken xori -1 form)
- CORRECTNESS-AUDIT-PASS2.md: findings 1-7 were already fixed in code
  but never marked; added per-finding Status: FIXED notes + banner
- regression/tests/float-nan-compare: new test covering all 13 FP
  compare predicates against NaN (f32+f64, runtime NaN); verifies the
  audit #1-3 NaN lowering fixes end-to-end. Suite: 62/62.

+++ tools/utilities/slow32dump.c
@@ -118,9 +118,9 @@ static bool dump_s32o(const char *filename, FILE *f, long file_size, options_t *
         return false;
     }
 
-    if (header.sec_offset + header.nsections * sizeof(s32o_section_t) > (uint32_t)file_size ||
-        header.sym_offset + header.nsymbols * sizeof(s32o_symbol_t) > (uint32_t)file_size ||
-        header.str_offset + header.str_size > (uint32_t)file_size) {
+    if ((uint64_t)header.sec_offset + (uint64_t)header.nsections * sizeof(s32o_section_t) > (uint64_t)file_size ||
+        (uint64_t)header.sym_offset + (uint64_t)header.nsymbols * sizeof(s32o_symbol_t) > (uint64_t)file_size ||
+        (uint64_t)header.str_offset + header.str_size > (uint64_t)file_size) {
         fprintf(stderr, "Error: Object file tables out of bounds\n");
         return false;
     }
@@ -231,7 +231,7 @@ static bool dump_s32o(const char *filename, FILE *f, long file_size, options_t *
         for (uint32_t i = 0; i < header.nsections; i++) {
             s32o_section_t *sec = &sections[i];
             if (sec->nrelocs == 0) continue;
-            if (sec->reloc_offset + sec->nrelocs * sizeof(s32o_reloc_t) > (uint32_t)file_size) {
+            if ((uint64_t)sec->reloc_offset + (uint64_t)sec->nrelocs * sizeof(s32o_reloc_t) > (uint64_t)file_size) {
                 fprintf(stderr, "Error: Relocation table out of bounds\n");
                 break;
             }
@@ -264,7 +264,7 @@ static bool dump_s32o(const char *filename, FILE *f, long file_size, options_t *
             s32o_section_t *sec = &sections[i];
             const char *name = (sec->name_offset < header.str_size) ? &strtab[sec->name_offset] : "<invalid>";
             if (sec->size == 0 || sec->type == S32_SEC_BSS) continue;
-            if (sec->offset + sec->size > (uint32_t)file_size) {
+            if ((uint64_t)sec->offset + sec->size > (uint64_t)file_size) {
                 fprintf(stderr, "Error: Section data out of bounds\n");
                 break;
             }
@@ -303,8 +303,8 @@ static bool dump_s32x(const char *filename, FILE *f, long file_size, options_t *
         return false;
     }
 
-    if (header.sec_offset + header.nsections * sizeof(s32x_section_t) > (uint32_t)file_size ||
-        header.str_offset + header.str_size > (uint32_t)file_size) {
+    if ((uint64_t)header.sec_offset + (uint64_t)header.nsections * sizeof(s32x_section_t) > (uint64_t)file_size ||
+        (uint64_t)header.str_offset + header.str_size > (uint64_t)file_size) {
         fprintf(stderr, "Error: Executable tables out of bounds\n");
         return false;
     }
